layers: Move synchronization CoreChecks to own file
